finish clean up, split sound mixer into platform folder
