+++ .vscode/settings.json
@@ -350,7 +350,8 @@
     "sound_mixer.h": "c",
     "cgb_audio.h": "c",
     "shared.h": "c",
-    "mp2k_common.h": "c"
+    "mp2k_common.h": "c",
+    "cgb_tables.h": "c"
   },
   "asm-code-lens.donated": true,
   "editor.formatOnSave": true,
