+++ include/lib/m4a/m4a_internal.h
@@ -2,10 +2,6 @@
 #define GUARD_GBA_M4A_INTERNAL_H
 
 #include "gba/gba.h"
-#include "lib/m4a/shared.h"
-#if PORTABLE
-#include "lib/m4a/music_player.h"
-#endif
 
 // ASCII encoding of 'Smsh' in reverse
 // This is presumably short for SMASH, the developer of MKS4AGB.
@@ -85,6 +81,187 @@
 #define FADE_IN        0x0002
 #define FADE_VOL_MAX   64
 #define FADE_VOL_SHIFT 2
+
+#define MAX_DIRECTSOUND_CHANNELS 12
+
+#if PLATFORM_GBA
+#define PCM_DMA_BUF_SIZE 1584 // size of Direct Sound buffer
+#else
+#define PCM_DMA_BUF_SIZE 4907 // size of Direct Sound buffer
+#endif
+
+#if PLATFORM_GBA
+#define NUM_MUSIC_PLAYERS ((u16)gNumMusicPlayers)
+#define MAX_LINES         ((u32)gMaxLines)
+#else
+#define NUM_MUSIC_PLAYERS 4
+#define MAX_LINES         0
+#endif
+
+struct MP2KTrack;
+struct MP2KPlayerState;
+
+typedef void (*MPlayFunc)();
+typedef void (*MP2KEventNoteFunc)(u8, struct MP2KPlayerState *, struct MP2KTrack *);
+typedef void (*MP2KEventFunc)(struct MP2KPlayerState *, struct MP2KTrack *);
+typedef void (*CgbSoundFunc)(void);
+typedef void (*CgbOscOffFunc)(u8);
+typedef u32 (*MidiKeyToCgbFreqFunc)(u8, u8, u8);
+typedef void (*ExtVolPitFunc)(void);
+typedef void (*MPlayMainFunc)(struct MP2KPlayerState *);
+
+struct MixerSource {
+    u8 status;
+    u8 type;
+    u8 rightVol;
+    u8 leftVol;
+
+    union {
+        struct {
+            u8 attack;
+            u8 decay;
+            u8 sustain;
+            u8 release;
+            u8 key;
+            u8 envelopeVol;
+
+            u8 envelopeGoal;
+            u8 envelopeCtr;
+
+            u8 echoVol;
+            u8 echoLen;
+            u8 padding1;
+            u8 padding2;
+            u8 gateTime;
+            u8 untransposedKey;
+            u8 velocity;
+            u8 priority;
+            u8 rhythmPan;
+            u8 padding3;
+            u8 padding4;
+            u8 padding5;
+
+            u8 padding6;
+            u8 sustainGoal;
+            u8 nrx4;
+            u8 pan;
+
+            u8 panMask;
+            u8 cgbStatus;
+            u8 length;
+            u8 sweep;
+
+            u32 freq;
+        } cgb;
+        struct {
+            u8 attack;
+            u8 decay;
+            u8 sustain;
+            u8 release;
+            u8 key;
+            u8 envelopeVol;
+
+            u8 envelopeVolR;
+            u8 envelopeVolL;
+
+            u8 echoVol;
+            u8 echoLen;
+            u8 padding1;
+            u8 padding2;
+            u8 gateTime;
+            u8 untransposedKey;
+            u8 velocity;
+            u8 priority;
+            u8 rhythmPan;
+            u8 padding3;
+            u8 padding4;
+            u8 padding5;
+
+            u32 ct;
+            float fw;
+
+            u32 freq;
+        } sound;
+    } data;
+
+    void *wav; // The next peice of data to be loading
+    void *current; // The working pointer from wav
+
+    struct MP2KTrack *track;
+    struct MixerSource *prev;
+    struct MixerSource *next;
+    u32 padding7; // d4
+    u32 blockCount; // bdpcm block count
+};
+
+struct SoundMixerState {
+    // This field is normally equal to ID_NUMBER but it is set to other
+    // values during sensitive operations for locking purposes.
+    u32 lockStatus;
+
+    vu8 dmaCounter;
+
+    // Direct Sound
+    u8 reverb;
+    u8 numChans;
+    u8 masterVol;
+    u8 freqOption;
+
+    u8 extensionFlags;
+    u8 cgbCounter15; // periodically counts from 14 down to 0 (15 states)
+    u8 framesPerDmaCycle; // number of V-blanks per PCM DMA
+    u8 maxScanlines;
+    u8 gap[3];
+    s32 samplesPerFrame;
+    s32 sampleRate;
+#if PLATFORM_GBA
+    s32 sampleRateReciprocal;
+#else
+    float sampleRateReciprocal;
+#endif
+    struct MixerSource *cgbChans;
+    MPlayMainFunc MPlayMainHead;
+    struct MP2KPlayerState *musicPlayerHead;
+    CgbSoundFunc CgbSound;
+    CgbOscOffFunc CgbOscOff;
+    MidiKeyToCgbFreqFunc MidiKeyToCgbFreq;
+    MPlayFunc *MPlayJumpTable;
+    MP2KEventNoteFunc plynote;
+    ExtVolPitFunc ExtVolPit;
+    void *reserved2;
+    void *reserved3;
+    void *reversed4;
+    void *reserved5;
+    struct MixerSource chans[MAX_DIRECTSOUND_CHANNELS];
+#if PLATFORM_GBA
+    s8 pcmBuffer[PCM_DMA_BUF_SIZE * 2];
+#else
+    // This is larger than needed for now to prevent
+    // some memory overflow which is occuring and messing
+    // with the globals after the sound info struct
+    float pcmBuffer[PCM_DMA_BUF_SIZE * 3];
+#endif
+};
+
+struct MP2KVoiceGroup {
+    u8 type;
+    u8 drumKey;
+    u8 cgbLength;
+    u8 pan_sweep; // pan or sweep (compatible sound ch. 1)
+    union {
+        struct {
+            struct WaveData *wav;
+            u8 attack;
+            u8 decay;
+            u8 sustain;
+            u8 release;
+        } sound;
+        struct {
+            struct MP2KVoiceGroup *group;
+            u8 *keySplitTable;
+        } keySplit;
+    } data;
+};
 
 struct WaveData {
     u16 type;
@@ -206,15 +383,11 @@ extern void *const gMPlayJumpTableTemplate[];
 extern const u8 gCgbScaleTable[];
 extern const s16 gCgbFreqTable[];
 extern const u8 gNoiseTable[];
+extern const u8 gCgb3Vol[];
 
-#if !PORTABLE
+#if PLATFORM_GBA
 extern char gNumMusicPlayers[];
 extern char gMaxLines[];
-#define NUM_MUSIC_PLAYERS ((u16)gNumMusicPlayers)
-#define MAX_LINES         ((u32)gMaxLines)
-#else
-#define NUM_MUSIC_PLAYERS 4
-#define MAX_LINES         0
 #endif
 
 u32 MidiKeyToFreq(struct WaveData *wav, u8 key, u8 fineAdjust);
